protect callbacks by a mutex
eventsexecutor is only a single thread. but the workaround to not queue
up timer events is to have a separate timer manager thread for the
events executor. To avoid parallel execution and possible race
conditions, every callback needs to be locked by a mutex to manually
achieve mutually exlusive callbacks.